orchestrator/requirements.txt
# Core FastAPI and async support
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis for caching and monitoring
redis[hiredis]==5.0.1

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.10.3
pydantic-settings==2.7.0

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# HTTP client
httpx>=0.27.0
aiohttp>=3.10.0
requests==2.31.0

# System utilities
psutil==5.9.6
docker==6.1.3

# Date and time
python-dateutil==2.8.2

# JSON processing
orjson==3.9.10

# Enhanced logging
python-json-logger==2.0.7
colorlog==6.7.0

# OpenAI and AI services
openai>=1.50.0
# dynaroute>=1.0.0  # Note: DynaRoute package not available in PyPI - install separately if needed

# Phoenix Arize AI Observability
arize-phoenix>=5.0.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-openai
opentelemetry-instrumentation-fastapi

# Machine Learning and NLP
numpy>=1.24.0
sentence-transformers>=2.2.0

# Security and content filtering
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
better-profanity==0.7.0

# Monitoring and metrics
prometheus-client>=0.17.0

# Web framework extensions
starlette>=0.27.0

# Configuration parsing
PyYAML>=6.0

# Command line interface (from Domain Identifier)
click>=8.0.0

# Development tools (optional)
# fastapi-cli  # Commented out to avoid typer conflicts

# spaCy language model (uncomment if needed)
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889